Favour VSCode over ESLint for import sorting
This silences the linter for import checks and replaces it with VSCode's built-in
organize-import, now implemented on save. It is less safe. But the two are not
compatible: VSCode sorts according to the import file, ESLint according to the
imported name. And it is to tedious to sort things manually.